+++ ios/Classes/CallkitIncomingAppDelegate.swift
@@ -6,16 +6,22 @@
 //
 
 import Foundation
+import AVFAudio
+import CallKit
 
 
 public protocol CallkitIncomingAppDelegate : NSObjectProtocol {
     
-    func onAccept(_ call: Call);
+    func onAccept(_ call: Call, _ action: CXAnswerCallAction);
     
     func onDecline(_ call: Call);
     
     func onEnd(_ call: Call);
     
     func onTimeOut(_ call: Call);
+
+    func didActivateAudioSession(_ audioSession: AVAudioSession)
+    
+    func didDeactivateAudioSession(_ audioSession: AVAudioSession)
     
 }